styles, format date, styles in the calendar
