+++ src/components/calendar/index.tsx
@@ -5,7 +5,7 @@ import { useGetVacanciesQuery } from "redux/VacancyQueries";
 import { useState } from "react";
 import ReminderItem from "components/reminder/ReminderItem";
 import "./calendar.css";
-import { colorVariants } from "components/notices/ShortNotice";
+import { colorVariants } from "utils/stylesHelpers";
 
 const CalendarComponent = () => {
   const [currentDay, setCurrentDay] = useState(new Date());
@@ -27,7 +27,7 @@ const CalendarComponent = () => {
   };
 
   return (
-    <ul className="lg:flex justify-between py-4 align-baseline">
+    <ul className="lg:flex justify-between py-4 align-baseline mb-28">
       <li className="w-[350px] mx-auto mb-8 lg:mb-0">
         <Calendar
           onChange={onChange}
@@ -44,14 +44,12 @@ const CalendarComponent = () => {
               const color = events.find(
                 ev => format(ev.actions[ev.actions.length - 1].deadline, "dd-MM-yyyy") === realDay
               )?.cardColor as string;
-              console.log("color-------------------", color);
               return `${colorVariants[color]} ${typeClass} rounded-full`;
             }
             if (format(date, "dd-MM-yyyy") === format(new Date(), "dd-MM-yyyy")) {
               return `bg-app-smoke text-txt-red ${typeClass} border border-1 border-bg-grey`;
             }
             return `${typeClass} rounded-full`;
-
           }}
         />
       </li>
