+++ src/components/notices/ShortNotice.tsx
@@ -4,6 +4,7 @@ import * as Icons from "components/iconsComponents";
 import Stars from "components/ui/stars";
 import { IVacancy } from "redux/VacancyQueries";
 import useHandleVacancy from "hooks/handleVacancy";
+import { colorVariants, effectIcon, effectItem } from "utils/stylesHelpers";
 
 type VacancyProps = {
   shortVacancy: IVacancy;
@@ -12,22 +13,6 @@ type VacancyProps = {
 export interface IColor {
   [key: string]: string;
 }
-
-export const colorVariants = {
-  red: "bg-app-red",
-  blue: "bg-app-blue",
-  green: "bg-app-green",
-  pink: "bg-app-pink",
-  smoke: "bg-app-smoke",
-  grey: "bg-app-grey",
-  yellow: "bg-app-yellow",
-  mustard: "bg-app-mustard",
-  orange: "bg-app-orange",
-} as IColor;
-
-export const effectIcon = `hover:scale-110 focus:scale-110 transition-transform duration-300`;
-
-export const effectItem = `hover:shadow-2xl focus:shadow-2xl`;
 
 const ShortNote = ({ shortVacancy }: VacancyProps) => {
   const { handleArchive, removeVacancy } = useHandleVacancy();
