+++ src/pages/Private/reminder.tsx
@@ -49,7 +49,7 @@ const Reminder = () => {
   return (
     <>
       <NavHeader prevAddress={location?.state?.from.pathname ?? "/"} bg="bg-bg-light" text="Reminder" underlined />
-      <div className="container mx-auto bg-bg-light">
+      <div className="container mx-auto bg-bg-light ">
         {isLoading ? (
           <div className="mt-48">
             <Loader active absolute />
